formatica-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
yt-dlp
pypdf==4.3.1
Pillow==10.4.0
aiofiles==23.2.1
httpx==0.27.2
python-magic==0.4.27
